Move blanket impls to not live under the ext feature
